Finish ColorPicker cleanup
