+++ FluentAvalonia/UI/Controls/ColorPicker/Enums/ColorComponent.cs
@@ -1,0 +1,13 @@
+﻿namespace FluentAvalonia.UI.Controls
+{
+	public enum ColorComponent
+    {
+        Hue,
+        Saturation,
+        Value,
+        Red,
+        Green,
+        Blue,
+        Alpha
+    }
+}
