+++ FluentAvalonia/UI/Controls/ColorPicker/Enums/ColorSpectrumShape.cs
@@ -1,0 +1,24 @@
+﻿namespace FluentAvalonia.UI.Controls
+{
+	/// <summary>
+	/// Defines values for how the <see cref="ColorSpectrum"/> should display
+	/// </summary>
+	public enum ColorSpectrumShape
+    {
+        /// <summary>
+        /// The <see cref="ColorSpectrum"/> should display a ColorWheel
+        /// </summary>
+        Wheel,
+
+        /// <summary>
+        /// The <see cref="ColorSpectrum"/> should display as a Color spectrum of two
+        /// color components
+        /// </summary>
+        Spectrum,
+
+        /// <summary>
+        /// The <see cref="ColorSpectrum"/> should display as a HSV Color Triangle
+        /// </summary>
+        Triangle
+    }
+}
